add navigation depending on cloud sync state
