+++ novawallet/Common/Services/CloudBackup/Model/CloudBackupSyncState.swift
@@ -4,4 +4,17 @@ enum CloudBackupSyncState: Equatable {
     case disabled(lastSyncDate: Date?)
     case unavailable(lastSyncDate: Date?)
     case enabled(CloudBackupSyncResult?, lastSyncDate: Date?)
+
+    var canAutoSync: Bool {
+        switch self {
+        case .disabled, .unavailable:
+            return false
+        case let .enabled(cloudBackupSyncResult, _):
+            if case .noUpdates = cloudBackupSyncResult {
+                return true
+            } else {
+                return false
+            }
+        }
+    }
 }
