+++ novawallet/Modules/WalletsList/Manage/WalletManageWireframe.swift
@@ -47,7 +47,13 @@ final class WalletManageWireframe: WalletBaseManageWireframe, WalletManageWirefr
         }
     }
 
-    func showImportWallet(from _: WalletManageViewProtocol?) {
-        // TODO: Create for adding
+    func showImportWallet(from view: WalletManageViewProtocol?) {
+        guard let importView = WalletImportOptionsViewFactory.createViewForAdding() else {
+            return
+        }
+
+        if let navigationController = view?.controller.navigationController {
+            navigationController.pushViewController(importView.controller, animated: true)
+        }
     }
 }
